+++ src/js/draw.js
@@ -32,7 +32,8 @@
  * @modified 2020-05-05 Added the 'lineWidth' param to the circle(...) function.
  * @modified 2020-05-12 Drawing any handles (square, circle, diamond) with lineWidth 1 now; this was not reset before.
  * @modified 2020-06-22 Added a context.clearRect() call to the clear() function; clearing with alpha channel did not work as expected.
- * @version  1.5.6
+ * @modified 2020-09-07 Added the circleArc(...) function to draw sections of circles.
+ * @version  1.6.0
  **/
 Object.defineProperty(exports, "__esModule", { value: true });
 var CubicBezierCurve_1 = require("./CubicBezierCurve");
@@ -383,6 +384,27 @@ var drawutils = /** @class */ (function () {
     };
     ;
     /**
+     * Draw a circular arc (section of a circle) with the given CSS color.
+     *
+     * @method circleArc
+     * @param {Vertex} center - The center of the circle.
+     * @param {number} radius - The radius of the circle.
+     * @param {number} startAngle - The angle to start at.
+     * @param {number} endAngle - The angle to end at.
+     * @param {string} color - The CSS color to draw the circle with.
+     * @return {void}
+     * @instance
+     * @memberof drawutils
+     */
+    drawutils.prototype.circleArc = function (center, radius, startAngle, endAngle, color, lineWidth) {
+        this.ctx.beginPath();
+        this.ctx.ellipse(this.offset.x + center.x * this.scale.x, this.offset.y + center.y * this.scale.y, radius * this.scale.x, radius * this.scale.y, 0.0, startAngle, endAngle, false);
+        this.ctx.closePath();
+        this.ctx.lineWidth = lineWidth || 1;
+        this._fillOrDraw(color);
+    };
+    ;
+    /**
      * Draw an ellipse with the specified (CSS-) color and thw two radii.
      *
      * @method ellipse
@@ -501,7 +523,7 @@ var drawutils = /** @class */ (function () {
     /**
      * Draw a diamond handle (square rotated by 45°) with the given CSS color.
      *
-     * It is an inherent featur of the handle functions that the drawn elements are not scaled and not
+     * It is an inherent feature of the handle functions that the drawn elements are not scaled and not
      * distorted. So even if the user zooms in or changes the aspect ratio, the handles will be drawn
      * as even shaped diamonds.
      *
@@ -527,7 +549,7 @@ var drawutils = /** @class */ (function () {
     /**
      * Draw a square handle with the given CSS color.<br>
      * <br>
-     * It is an inherent featur of the handle functions that the drawn elements are not scaled and not
+     * It is an inherent feature of the handle functions that the drawn elements are not scaled and not
      * distorted. So even if the user zooms in or changes the aspect ratio, the handles will be drawn
      * as even shaped squares.
      *
@@ -550,7 +572,7 @@ var drawutils = /** @class */ (function () {
     /**
      * Draw a circle handle with the given CSS color.<br>
      * <br>
-     * It is an inherent featur of the handle functions that the drawn elements are not scaled and not
+     * It is an inherent feature of the handle functions that the drawn elements are not scaled and not
      * distorted. So even if the user zooms in or changes the aspect ratio, the handles will be drawn
      * as even shaped circles.
      *