+++ src/ts/VertTuple.ts
@@ -164,8 +164,9 @@ export class VertTuple<T extends VertTuple<T>> {
      * @memberof VertTuple
      **/
     angle( line?:VertTuple<any> ):number {
-	if( typeof line == 'undefined' )
+	if( line == null || typeof line == 'undefined' ) {
 	    line = this.factory( new Vertex(0,0), new Vertex(100,0) );
+	}
 	// Compute the angle from x axis and the return the difference :)
 	var v0 = this.b.clone().sub( this.a );
 	var v1 = line.b.clone().sub( line.a );